add Makefile

+++ src/driver/mod.rs
@@ -1,15 +1,16 @@
 use core::ptr::{read, write};
 
 
-
-pub fn mmio_read(addr: usize) -> usize {
+/// Read something from mmio address
+pub fn mmio_read(addr: u32) -> u32 {
     unsafe{
-        let val = read(addr as *const usize);
+        let val = read(addr as *const u32);
     }
 }
 
-pub fn mmio_write(addr: usize, val: usize) {
+/// Write something to mmio address
+pub fn mmio_write(addr: u32, val: u32) {
     unsafe{
-        write(addr as *mut usize, val);
+        write(addr as *mut u32, val);
     }
 }
